Refactoring in preparation for FastBioNLP.
